Add Redirect feature to server
Create a redirect feature that responds with a 302 status along with the location header.
- Redirect Handler
- Redirect Route
- Redirect Cukes
