+++ src/main/java/Config/Routes/Routes.java
@@ -6,17 +6,17 @@ import Request.Request;
 import java.util.ArrayList;
 
 public class Routes {
-  public ArrayList<Route> routes = new ArrayList<>();
+  public static ArrayList<Route> routes = new ArrayList<>();
 
   public void add(Route route) {
     routes.add(route);
   }
 
-  public ArrayList<Route> getRoutes() {
+  public static ArrayList<Route> getRoutes() {
     return routes;
   }
 
-  private Route getNotFound() {
+  public static Route getNotFound() {
     Route notFound = null;
     for (Route route : getRoutes()) {
       if (route.method.equals(Method.GET) && route.path.equals("/not-found")) notFound = route;
