+++ src/main/java/Controller/Handler/GETEchoHandler.java
@@ -7,8 +7,7 @@ public class GETEchoHandler extends Handler {
   public Response handle(Request request) {
     return new Response.Builder()
       .setStatusCode(200)
-      .setHeader("Content-Length", "0")
-      .setHeader("Content-Type", "text/html")
+      .setBody("")
       .build();
   }
 }
