+++ src/main/java/Controller/Handler/RedirectHandler.java
@@ -2,11 +2,12 @@ package Controller.Handler;
 import Request.Request;
 import Response.Response;
 
-public class GETHandler extends Handler {
+public class RedirectHandler extends Handler {
   public Response handle(Request request) {
+
     return new Response.Builder()
-      .setStatusCode(200)
-      .setBody("")
+      .setStatusCode(302)
+      .setHeader("Location", "http://127.0.0.1:3000/")
       .build();
   }
 }
