+++ src/test/java/Config/Routes/RouteTest.java
@@ -2,12 +2,11 @@ package Config.Routes;
 
 import Controller.Handler.GETHandler;
 import Controller.Handler.POSTEchoHandler;
-import Request.Request;
 import org.junit.Test;
 
 import static Config.Method.GET;
 import static Config.Method.POST;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 public class RouteTest {
   private String CRLF = "\r\n";
@@ -25,35 +24,6 @@ public class RouteTest {
   }
 
   @Test
-  public void testGETHandlerResponseStatusCode() {
-    Route route = new Route(GET, "/", new GETHandler());
-    String requestString = "GET / HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF;
-    Request request = new Request.Builder().build(requestString);
-    assertEquals(route.handler.handle(request).statusCode, 200);
-  }
-
-  @Test
-  public void testGETHandlerResponseContentLength() {
-    Route route = new Route(GET, "/", new GETHandler());
-    String requestString = "GET / HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF;
-    Request request = new Request.Builder().build(requestString);
-    assertEquals(route.handler.handle(request).contentLength, 0);
-  }
-
-  @Test
-  public void testGETHandlerResponseBody() {
-    Route route = new Route(GET, "/", new GETHandler());
-    String requestString = "GET / HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF;
-    Request request = new Request.Builder().build(requestString);
-    assertEquals(route.handler.handle(request).body, "");
-  }
-  @Test
   public void testPOSTMethod() {
     Route route = new Route(POST, "/echo", new POSTEchoHandler());
     assertEquals(POST, route.method);
@@ -64,38 +34,4 @@ public class RouteTest {
     Route route = new Route(POST, "/echo", new POSTEchoHandler());
     assertEquals("/echo", route.path);
   }
-
-  @Test
-  public void testPOSTEchoHandlerResponseStatusCode() {
-    Route route = new Route(POST, "/echo", new POSTEchoHandler());
-    String requestString = "POST /echo HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF +
-      "goodbye";
-    Request request = new Request.Builder().build(requestString);
-    assertEquals(200, route.handler.handle(request).statusCode);
-  }
-
-  @Test
-  public void testPOSTEchoHandlerResponseContentLength() {
-    Route route = new Route(POST, "/echo", new POSTEchoHandler());
-    String requestString = "POST /echo HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF +
-      "goodbye";
-    Request request = new Request.Builder().build(requestString);
-    assertEquals(7, route.handler.handle(request).contentLength);
-  }
-
-  @Test
-  public void testPOSTEchoHandlerResponseBody() {
-    Route route = new Route(POST, "/echo", new POSTEchoHandler());
-    String requestString = "POST /echo HTTP/1.1" + CRLF +
-      "Content-Type: text/plain" + CRLF +
-      "content-length: 7" + CRLF + CRLF +
-      "goodbye";
-    Request request = new Request.Builder().build(requestString);
-    assertEquals("goodbye", route.handler.handle(request).body);
-  }
-
 }
